Reorganise the repository
Despite a source directory making grepping easier, I think a flat structure is easier to work with in many other ways.

Page names were automatically generated from old manual, but were not concise and in many cases redundant. Removing redundancy, "the"/"and", "panda3d", etc. in cases where it does not harm clarity is what I used as a general rule of thumb.